requirements.txt
openai==0.27.0
gradio==3.20.1
python-dotenv==0.21.1
pytube==12.1.2
pydub==0.25.1
youtube_transcript_api==0.5.0